feat(edit): remove leftover JSON and DTO helpers from live parameter access
Live GPU packing writes packed structs, the live snapshot no longer dumps stage JSON, and JSON remains only at serialization boundaries.
