+++ alcedo_studio/src/edit/runtime/metal/metal_develop_pass.mm
@@ -27,8 +27,6 @@
 #include "edit/geometry/types.hpp"
 #include "edit/graph/develop_color_transform.hpp"
 #include "edit/graph/graph_ids.hpp"
-#include "edit/operators/models/builtin_type_ids.hpp"
-#include "edit/operators/models/operator_param_dto.hpp"
 #include "edit/operators/models/pending_parameter_patch.hpp"
 #include "edit/runtime/camera_color_gpu_params.hpp"
 #include "edit/runtime/develop_demosaic.hpp"
@@ -360,7 +358,7 @@ void ExecuteMetalDevelop(MetalRenderDevice& device, const ExecutionPlan& plan,
   if (develop == nullptr) {
     throw std::runtime_error("ExecuteMetalDevelop: missing develop node");
   }
-  auto pending = TakePendingParameterPatch(develop->Params());
+  auto pending = TakePendingDirtyFields(develop->Params());
   if (workspace.Textures().ByteBudget() == 0) {
     workspace.Textures().SetByteBudget(MetalBackend::DefaultTextureBudgetBytes());
   }
@@ -506,20 +504,9 @@ void ExecuteMetalCameraColor(MetalRenderDevice& device, const ExecutionPlan& pla
   for (int i = 0; i < 9; ++i) {
     gpu_params.camera_to_ap1[i] = resolved.transform.camera_to_ap1[static_cast<std::size_t>(i)];
   }
-  auto&                       arena = workspace.Parameters();
-  const ParameterSlotKey      key{develop->Id(), kDevelopCameraColorSlot};
-  const ParameterFieldBinding field{DirtyFieldMask{DevelopDirty::WhiteBalance}, 0, 0,
-                                    sizeof(CameraColorGpuParams)};
-  auto payload = std::make_shared<TypedOperatorParamPayload<CameraColorGpuParams>>(
-      type_ids::DevelopNode(), 1, gpu_params);
-  if (!arena.Contains(key)) {
-    arena.BindSlot(key, sizeof(CameraColorGpuParams), std::span{&field, 1});
-    arena.InitializeFromFullDto(key, OperatorParamDto{type_ids::DevelopNode(), 1, payload});
-  } else {
-    OperatorParamPatchDto patch{develop->Id(), kDevelopCameraColorSlot, type_ids::DevelopNode(),
-                                DirtyFieldMask{DevelopDirty::WhiteBalance}, payload};
-    arena.ApplyPatch(key, patch);
-  }
+  auto&                  arena = workspace.Parameters();
+  const ParameterSlotKey key{develop->Id(), kDevelopCameraColorSlot};
+  arena.BindOrWritePackedSlot(key, DirtyFieldMask{DevelopDirty::WhiteBalance}, gpu_params);
   arena.UploadDirty(device.CommandContext());
   const auto binding        = arena.Binding(key);
   auto*      command_buffer = CommandBuffer(device);
