+++ alcedo_studio/src/edit/runtime/metal/metal_drt_pass.mm
@@ -240,28 +240,18 @@ auto ExecuteMetalDrt(MetalRenderDevice& device, const ExecutionPlan& plan,
     }
   }
 
-  const ParameterSlotKey      key{drt->Id(), AdjustmentInstanceId{"drt.output"}};
-  const ParameterFieldBinding field{DirtyFieldMask{kDrtDirtyBits}, 0, 0, sizeof(MetalDrtGpuParams)};
-  auto                        pending          = plan.output_color_override.has_value()
-                                ? decltype(TakePendingParameterPatch(drt->Params())){}
-                                : TakePendingParameterPatch(drt->Params());
-  const bool                  needs_initialize = !arena.Contains(key);
+  const ParameterSlotKey key{drt->Id(), AdjustmentInstanceId{"drt.output"}};
+  auto                   pending = plan.output_color_override.has_value()
+                                       ? decltype(TakePendingDirtyFields(drt->Params())){}
+                                       : TakePendingDirtyFields(drt->Params());
+  const bool             needs_initialize = !arena.Contains(key);
   if (needs_initialize || pending.has_value() || plan.output_color_override.has_value()) {
     auto drt_json = drt->Params().ToJson();
     if (plan.output_color_override.has_value()) {
       OverlayExportColorOnDrtJson(drt_json, *plan.output_color_override);
     }
     const auto runtime = ResolveMetalDrtGpuParams(drt_json);
-    auto       payload = std::make_shared<TypedOperatorParamPayload<MetalDrtGpuParams>>(
-        drt->Params().Type(), 1, runtime);
-    if (needs_initialize) {
-      arena.BindSlot(key, sizeof(MetalDrtGpuParams), std::span{&field, 1});
-      arena.InitializeFromFullDto(key, OperatorParamDto{drt->Params().Type(), 1, payload});
-    } else {
-      arena.ApplyPatch(key, OperatorParamPatchDto{drt->Id(), AdjustmentInstanceId{"drt.output"},
-                                                 drt->Params().Type(), DirtyFieldMask{kDrtDirtyBits},
-                                                 payload});
-    }
+    arena.BindOrWritePackedSlot(key, DirtyFieldMask{kDrtDirtyBits}, runtime);
   }
   arena.UploadDirty(context);
   if (pending) {
