+++ alcedo_studio/tests/edit/runtime/metal_grade_test.cpp
@@ -356,6 +356,23 @@ TEST_F(MetalGradeFixture, MetalGradeParameterBindDoesNotCopyFullDto) {
   EXPECT_EQ(OperatorModelFullDtoCopyCount::Peek(), 0);
 }
 
+TEST_F(MetalGradeFixture, MetalCameraColorPackedWriteDoesNotCopyFullDto) {
+  device_.BeginRender();
+  ExecuteMetalDevelop(device_, plan_, prepared_, document_);
+  ExecuteMetalGeometryResample(device_, plan_);
+  OperatorModelFullDtoCopyCount::Reset();
+  ExecuteMetalCameraColor(device_, plan_, document_);
+  device_.EndRender();
+  device_.WaitIdle();
+  EXPECT_EQ(OperatorModelFullDtoCopyCount::Peek(), 0);
+}
+
+TEST_F(MetalGradeFixture, MetalDrtPackedWriteDoesNotCopyFullDto) {
+  OperatorModelFullDtoCopyCount::Reset();
+  (void)RenderThroughDrtPost();
+  EXPECT_EQ(OperatorModelFullDtoCopyCount::Peek(), 0);
+}
+
 TEST_F(MetalGradeFixture, MetalExposureEditRunsOnlyPrimaryGradeAndDrt) {
   (void)device_.Execute(plan_, prepared_, document_);
   device_.WaitIdle();
