+++ alcedo_studio/tests/edit/runtime/metal_workspace_test.cpp
@@ -20,30 +20,30 @@ namespace {
 using metal_workspace_test::BindSharpen;
 using metal_workspace_test::ExposureFieldBindings;
 using metal_workspace_test::MetalWorkspaceFixture;
-using metal_workspace_test::UploadFullAndClearDirty;
+using metal_workspace_test::UploadPackedAndClearDirty;
+using metal_workspace_test::WritePackedOwnerBytes;
 
-TEST_F(MetalWorkspaceFixture, MetalParameterArenaUploadsOnlyDirtyRanges) {
+TEST_F(MetalWorkspaceFixture, MetalParameterArenaWritePackedSlotUploadsBoundSlotOnce) {
   MetalRenderDevice device;
   ParameterSlotKey  key{NodeId{"grade.primary"}, AdjustmentInstanceId{"sharpen"}};
   SharpenModel      model;
   BindSharpen(device.Workspace().Parameters(), key);
-  ASSERT_TRUE(UploadFullAndClearDirty(device, key, model));
+  ASSERT_TRUE(UploadPackedAndClearDirty(device, key, model));
 
   model.SetAmount(12.0f);
-  auto pending = TakePendingParameterPatch(model);
+  auto pending = TakePendingDirtyFields(model);
   ASSERT_TRUE(pending.has_value());
 
   auto& backend = device.Workspace().Device();
   backend.ResetCounters();
-  device.Workspace().Parameters().ApplyPatch(key, pending->Patch());
+  WritePackedOwnerBytes(device.Workspace().Parameters(), key, model);
   device.Workspace().Parameters().UploadDirty(device.CommandContext());
   device.WaitIdle();
   pending->Commit();
 
   ASSERT_EQ(backend.LastHostToDeviceRanges().size(), 1U);
-  EXPECT_EQ(backend.LastHostToDeviceRanges().front().size, 4U);
-  EXPECT_EQ(backend.HostToDeviceBytes(), 4U);
-  EXPECT_LT(backend.HostToDeviceBytes(), sizeof(SharpenPayload));
+  EXPECT_EQ(backend.LastHostToDeviceRanges().front().size, sizeof(SharpenPayload));
+  EXPECT_EQ(backend.HostToDeviceBytes(), sizeof(SharpenPayload));
 
   backend.ResetCounters();
   device.Workspace().Parameters().UploadDirty(device.CommandContext());
@@ -245,14 +245,14 @@ TEST_F(MetalWorkspaceFixture, MetalFailedUploadRestoresDirtyFieldsAndPublishesNo
   ExposureModel     model;
   const auto        fields = ExposureFieldBindings();
   device.Workspace().Parameters().BindSlot(key, 4, fields);
-  ASSERT_TRUE(UploadFullAndClearDirty(device, key, model));
+  ASSERT_TRUE(UploadPackedAndClearDirty(device, key, model));
 
   model.SetValue(0.75f);
   {
-    auto pending = TakePendingParameterPatch(model);
+    auto pending = TakePendingDirtyFields(model);
     ASSERT_TRUE(pending.has_value());
     device.Workspace().Device().FailNextUpload();
-    device.Workspace().Parameters().ApplyPatch(key, pending->Patch());
+    WritePackedOwnerBytes(device.Workspace().Parameters(), key, model);
     EXPECT_THROW(device.Workspace().Parameters().UploadDirty(device.CommandContext()),
                  std::runtime_error);
   }
@@ -292,13 +292,13 @@ TEST_F(MetalWorkspaceFixture, ParameterUploadFailureRestoresPendingDirtyState) {
   ExposureModel     model;
   const auto        fields = ExposureFieldBindings();
   device.Workspace().Parameters().BindSlot(key, 4, fields);
-  ASSERT_TRUE(UploadFullAndClearDirty(device, key, model));
+  ASSERT_TRUE(UploadPackedAndClearDirty(device, key, model));
 
   model.SetValue(0.75f);
-  auto pending = TakePendingParameterPatch(model);
+  auto pending = TakePendingDirtyFields(model);
   ASSERT_TRUE(pending.has_value());
   device.Workspace().Device().FailNextUpload();
-  device.Workspace().Parameters().ApplyPatch(key, pending->Patch());
+  WritePackedOwnerBytes(device.Workspace().Parameters(), key, model);
   EXPECT_TRUE(device.Workspace().Parameters().HasPendingUpload());
   EXPECT_THROW(device.Workspace().Parameters().UploadDirty(device.CommandContext()),
                std::runtime_error);
